Split getCallInfo RyuJit and CPAOT implementations
The overall structure of the getCallInfo implementation is starting to diverge from what we have in jitinterface.cpp for Crossgen. To aid the bring-up of CPAOT and reduce bugs from the complex logic getting re-ordered, I think it's helpful to split the implementation so we can remove the various #if checks. This frees us to match the ready-to-run implementation more closely with Crossgen. This is helpful for example with the ldvirtftn fix, where the runtime failure was caused by our getCallInfo implementation overwriting a previously set value for `pResult->codePointerOrStubLookup`, something that Crossgen's jitinterface avoids.
