+++ src/ILCompiler.RyuJit/src/JitInterface/CorInfoImpl.RyuJit.cs
@@ -5,6 +5,7 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Runtime.InteropServices;
 
 using Internal.IL;
 using Internal.TypeSystem;
@@ -955,6 +956,374 @@ namespace Internal.JitInterface
                 return CorInfoHelpFunc.CORINFO_HELP_NEWARR_1_ALIGN8;
 
             return CorInfoHelpFunc.CORINFO_HELP_NEWARR_1_VC;
+        }
+
+        private void getCallInfo(ref CORINFO_RESOLVED_TOKEN pResolvedToken, CORINFO_RESOLVED_TOKEN* pConstrainedResolvedToken, CORINFO_METHOD_STRUCT_* callerHandle, CORINFO_CALLINFO_FLAGS flags, CORINFO_CALL_INFO* pResult)
+        {
+#if DEBUG
+            // In debug, write some bogus data to the struct to ensure we have filled everything
+            // properly.
+            MemoryHelper.FillMemory((byte*)pResult, 0xcc, Marshal.SizeOf<CORINFO_CALL_INFO>());
+#endif
+            MethodDesc method = HandleToObject(pResolvedToken.hMethod);
+
+            // Spec says that a callvirt lookup ignores static methods. Since static methods
+            // can't have the exact same signature as instance methods, a lookup that found
+            // a static method would have never found an instance method.
+            if (method.Signature.IsStatic && (flags & CORINFO_CALLINFO_FLAGS.CORINFO_CALLINFO_CALLVIRT) != 0)
+            {
+                throw new BadImageFormatException();
+            }
+
+            // This block enforces the rule that methods with [NativeCallable] attribute
+            // can only be called from unmanaged code. The call from managed code is replaced
+            // with a stub that throws an InvalidProgramException
+            if (method.IsNativeCallable && (flags & CORINFO_CALLINFO_FLAGS.CORINFO_CALLINFO_LDFTN) == 0)
+            {
+                ThrowHelper.ThrowInvalidProgramException(ExceptionStringID.InvalidProgramNativeCallable, method);
+            }
+
+            TypeDesc exactType = HandleToObject(pResolvedToken.hClass);
+
+            TypeDesc constrainedType = null;
+            if ((flags & CORINFO_CALLINFO_FLAGS.CORINFO_CALLINFO_CALLVIRT) != 0 && pConstrainedResolvedToken != null)
+            {
+                constrainedType = HandleToObject(pConstrainedResolvedToken->hClass);
+            }
+
+            bool resolvedConstraint = false;
+            bool forceUseRuntimeLookup = false;
+            bool targetIsFatFunctionPointer = false;
+
+            MethodDesc methodAfterConstraintResolution = method;
+            if (constrainedType == null)
+            {
+                pResult->thisTransform = CORINFO_THIS_TRANSFORM.CORINFO_NO_THIS_TRANSFORM;
+            }
+            else
+            {
+                // We have a "constrained." call.  Try a partial resolve of the constraint call.  Note that this
+                // will not necessarily resolve the call exactly, since we might be compiling
+                // shared generic code - it may just resolve it to a candidate suitable for
+                // JIT compilation, and require a runtime lookup for the actual code pointer
+                // to call.
+
+                MethodDesc directMethod = constrainedType.GetClosestDefType().TryResolveConstraintMethodApprox(exactType, method, out forceUseRuntimeLookup);
+                if (directMethod == null && constrainedType.IsEnum)
+                {
+                    // Constrained calls to methods on enum methods resolve to System.Enum's methods. System.Enum is a reference
+                    // type though, so we would fail to resolve and box. We have a special path for those to avoid boxing.
+                    directMethod = _compilation.TypeSystemContext.TryResolveConstrainedEnumMethod(constrainedType, method);
+                }
+
+                if (directMethod != null)
+                {
+                    // Either
+                    //    1. no constraint resolution at compile time (!directMethod)
+                    // OR 2. no code sharing lookup in call
+                    // OR 3. we have have resolved to an instantiating stub
+
+                    methodAfterConstraintResolution = directMethod;
+
+                    Debug.Assert(!methodAfterConstraintResolution.OwningType.IsInterface);
+                    resolvedConstraint = true;
+                    pResult->thisTransform = CORINFO_THIS_TRANSFORM.CORINFO_NO_THIS_TRANSFORM;
+
+                    exactType = constrainedType;
+                }
+                else if (constrainedType.IsValueType)
+                {
+                    pResult->thisTransform = CORINFO_THIS_TRANSFORM.CORINFO_BOX_THIS;
+                }
+                else
+                {
+                    pResult->thisTransform = CORINFO_THIS_TRANSFORM.CORINFO_DEREF_THIS;
+                }
+            }
+
+            MethodDesc targetMethod = methodAfterConstraintResolution;
+
+            //
+            // Initialize callee context used for inlining and instantiation arguments
+            //
+
+
+            if (targetMethod.HasInstantiation)
+            {
+                pResult->contextHandle = contextFromMethod(targetMethod);
+                pResult->exactContextNeedsRuntimeLookup = targetMethod.IsSharedByGenericInstantiations;
+            }
+            else
+            {
+                pResult->contextHandle = contextFromType(exactType);
+                pResult->exactContextNeedsRuntimeLookup = exactType.IsCanonicalSubtype(CanonicalFormKind.Any);
+            }
+
+            //
+            // Determine whether to perform direct call
+            //
+
+            bool directCall = false;
+            bool resolvedCallVirt = false;
+
+            if (targetMethod.Signature.IsStatic)
+            {
+                // Static methods are always direct calls
+                directCall = true;
+            }
+            else if (targetMethod.OwningType.IsInterface)
+            {
+                // Force all interface calls to be interpreted as if they are virtual.
+                directCall = false;
+            }
+            else if ((flags & CORINFO_CALLINFO_FLAGS.CORINFO_CALLINFO_CALLVIRT) == 0 || resolvedConstraint)
+            {
+                directCall = true;
+            }
+            else
+            {
+                if (!targetMethod.IsVirtual || targetMethod.IsFinal || targetMethod.OwningType.IsSealed())
+                {
+                    resolvedCallVirt = true;
+                    directCall = true;
+                }
+            }
+
+            pResult->codePointerOrStubLookup.lookupKind.needsRuntimeLookup = false;
+
+            bool allowInstParam = (flags & CORINFO_CALLINFO_FLAGS.CORINFO_CALLINFO_ALLOWINSTPARAM) != 0;
+
+            if (directCall && !allowInstParam && targetMethod.GetCanonMethodTarget(CanonicalFormKind.Specific).RequiresInstArg())
+            {
+                // JIT needs a single address to call this method but the method needs a hidden argument.
+                // We need a fat function pointer for this that captures both things.
+                targetIsFatFunctionPointer = true;
+
+                // JIT won't expect fat function pointers unless this is e.g. delegate creation
+                Debug.Assert((flags & CORINFO_CALLINFO_FLAGS.CORINFO_CALLINFO_LDFTN) != 0);
+
+                pResult->kind = CORINFO_CALL_KIND.CORINFO_CALL_CODE_POINTER;
+
+                if (pResult->exactContextNeedsRuntimeLookup)
+                {
+                    pResult->codePointerOrStubLookup.lookupKind.needsRuntimeLookup = true;
+                    pResult->codePointerOrStubLookup.lookupKind.runtimeLookupFlags = 0;
+                    pResult->codePointerOrStubLookup.runtimeLookup.indirections = CORINFO.USEHELPER;
+
+                    // Do not bother computing the runtime lookup if we are inlining. The JIT is going
+                    // to abort the inlining attempt anyway.
+                    MethodDesc contextMethod = methodFromContext(pResolvedToken.tokenContext);
+                    if (contextMethod == MethodBeingCompiled)
+                    {
+                        pResult->codePointerOrStubLookup.lookupKind.runtimeLookupKind = GetGenericRuntimeLookupKind(contextMethod);
+                        pResult->codePointerOrStubLookup.lookupKind.runtimeLookupFlags = (ushort)ReadyToRunHelperId.MethodEntry;
+                        pResult->codePointerOrStubLookup.lookupKind.runtimeLookupArgs = (void*)ObjectToHandle(GetRuntimeDeterminedObjectForToken(ref pResolvedToken));
+                    }
+                }
+                else
+                {
+                    pResult->codePointerOrStubLookup.constLookup =
+                        CreateConstLookupToSymbol(_compilation.NodeFactory.FatFunctionPointer(targetMethod));
+                }
+            }
+            else if (directCall)
+            {
+                bool referencingArrayAddressMethod = false;
+
+                if (targetMethod.IsIntrinsic)
+                {
+                    // If this is an intrinsic method with a callsite-specific expansion, this will replace
+                    // the method with a method the intrinsic expands into. If it's not the special intrinsic,
+                    // method stays unchanged.
+                    var methodIL = (MethodIL)HandleToObject((IntPtr)pResolvedToken.tokenScope);
+                    targetMethod = _compilation.ExpandIntrinsicForCallsite(targetMethod, methodIL.OwningMethod);
+
+                    // For multidim array Address method, we pretend the method requires a hidden instantiation argument
+                    // (even though it doesn't need one). We'll actually swap the method out for a differnt one with
+                    // a matching calling convention later. See ArrayMethod for a description.
+                    referencingArrayAddressMethod = targetMethod.IsArrayAddressMethod();
+                }
+
+                MethodDesc concreteMethod = targetMethod;
+                targetMethod = targetMethod.GetCanonMethodTarget(CanonicalFormKind.Specific);
+
+                pResult->kind = CORINFO_CALL_KIND.CORINFO_CALL;
+
+                if (targetMethod.IsConstructor && targetMethod.OwningType.IsString)
+                {
+                    // Calling a string constructor doesn't call the actual constructor.
+                    pResult->codePointerOrStubLookup.constLookup = CreateConstLookupToSymbol(
+                        _compilation.NodeFactory.StringAllocator(targetMethod)
+                        );
+                }
+                else if (pResult->exactContextNeedsRuntimeLookup)
+                {
+                    // Nothing to do... The generic handle lookup gets embedded in to the codegen
+                    // during the jitting of the call.
+                    // (Note: The generic lookup in R2R is performed by a call to a helper at runtime, not by
+                    // codegen emitted at crossgen time)
+
+                    Debug.Assert(!forceUseRuntimeLookup);
+                    pResult->codePointerOrStubLookup.constLookup = CreateConstLookupToSymbol(
+                        _compilation.NodeFactory.MethodEntrypoint(targetMethod)
+                        );
+                }
+                else
+                {
+                    ISymbolNode instParam = null;
+
+                    if (targetMethod.RequiresInstMethodDescArg())
+                    {
+                        instParam = _compilation.NodeFactory.MethodGenericDictionary(concreteMethod);
+                    }
+                    else if (targetMethod.RequiresInstMethodTableArg() || referencingArrayAddressMethod)
+                    {
+                        // Ask for a constructed type symbol because we need the vtable to get to the dictionary
+                        instParam = _compilation.NodeFactory.ConstructedTypeSymbol(concreteMethod.OwningType);
+                    }
+
+                    if (instParam != null)
+                    {
+                        pResult->instParamLookup = CreateConstLookupToSymbol(instParam);
+                    }
+
+                    pResult->codePointerOrStubLookup.constLookup = CreateConstLookupToSymbol(
+                        _compilation.NodeFactory.MethodEntrypoint(targetMethod)
+                        );
+                }
+
+                pResult->nullInstanceCheck = resolvedCallVirt;
+            }
+            else if (method.HasInstantiation)
+            {
+                // GVM Call Support
+                pResult->kind = CORINFO_CALL_KIND.CORINFO_VIRTUALCALL_LDVIRTFTN;
+                pResult->codePointerOrStubLookup.constLookup.accessType = InfoAccessType.IAT_VALUE;
+                pResult->nullInstanceCheck = true;
+
+                if (pResult->exactContextNeedsRuntimeLookup)
+                {
+                    ComputeLookup(ref pResolvedToken,
+                        GetRuntimeDeterminedObjectForToken(ref pResolvedToken),
+                        ReadyToRunHelperId.MethodHandle,
+                        ref pResult->codePointerOrStubLookup);
+                    Debug.Assert(pResult->codePointerOrStubLookup.lookupKind.needsRuntimeLookup);
+                }
+
+                // RyuJIT will assert if we report CORINFO_CALLCONV_PARAMTYPE for a result of a ldvirtftn
+                // We don't need an instantiation parameter, so let's just not report it. Might be nice to
+                // move that assert to some place later though.
+                targetIsFatFunctionPointer = true;
+            }
+            else
+            // In ReadyToRun, we always use the dispatch stub to call virtual methods
+            if ((flags & CORINFO_CALLINFO_FLAGS.CORINFO_CALLINFO_LDFTN) == 0
+                && targetMethod.OwningType.IsInterface)
+            {
+                pResult->kind = CORINFO_CALL_KIND.CORINFO_VIRTUALCALL_STUB;
+
+                if (pResult->exactContextNeedsRuntimeLookup)
+                {
+                    ComputeLookup(ref pResolvedToken,
+                        GetRuntimeDeterminedObjectForToken(ref pResolvedToken),
+                        ReadyToRunHelperId.VirtualDispatchCell,
+                        ref pResult->codePointerOrStubLookup);
+                    Debug.Assert(pResult->codePointerOrStubLookup.lookupKind.needsRuntimeLookup);
+                }
+                else
+                {
+                    pResult->codePointerOrStubLookup.lookupKind.needsRuntimeLookup = false;
+                    pResult->codePointerOrStubLookup.constLookup.accessType = InfoAccessType.IAT_PVALUE;
+                    pResult->codePointerOrStubLookup.constLookup.addr = (void*)ObjectToHandle(
+                        _compilation.NodeFactory.InterfaceDispatchCell(targetMethod
+#if !SUPPORT_JIT
+                        , _compilation.NameMangler.GetMangledMethodName(MethodBeingCompiled).ToString()
+#endif
+                        ));
+                }
+            }
+            else
+            if ((flags & CORINFO_CALLINFO_FLAGS.CORINFO_CALLINFO_LDFTN) == 0
+                && _compilation.HasFixedSlotVTable(targetMethod.OwningType))
+            {
+                pResult->kind = CORINFO_CALL_KIND.CORINFO_VIRTUALCALL_VTABLE;
+                pResult->nullInstanceCheck = true;
+            }
+            else
+            {
+                ReadyToRunHelperId helperId;
+                if ((flags & CORINFO_CALLINFO_FLAGS.CORINFO_CALLINFO_LDFTN) != 0)
+                {
+                    pResult->kind = CORINFO_CALL_KIND.CORINFO_VIRTUALCALL_LDVIRTFTN;
+                    helperId = ReadyToRunHelperId.ResolveVirtualFunction;
+                }
+                else
+                {
+                    // CORINFO_CALL_CODE_POINTER tells the JIT that this is indirect
+                    // call that should not be inlined.
+                    pResult->kind = CORINFO_CALL_KIND.CORINFO_CALL_CODE_POINTER;
+                    helperId = ReadyToRunHelperId.VirtualCall;
+                }
+
+                // If this is a non-interface call, we actually don't need a runtime lookup to find the target.
+                // We don't even need to keep track of the runtime-determined method being called because the system ensures
+                // that if e.g. Foo<__Canon>.GetHashCode is needed and we're generating a dictionary for Foo<string>,
+                // Foo<string>.GetHashCode is needed too.
+                if (pResult->exactContextNeedsRuntimeLookup && targetMethod.OwningType.IsInterface)
+                {
+                    // We need JitInterface changes to fully support this.
+                    // If this is LDVIRTFTN of an interface method that is part of a verifiable delegate creation sequence,
+                    // RyuJIT is not going to use this value.
+                    Debug.Assert(helperId == ReadyToRunHelperId.ResolveVirtualFunction);
+                    pResult->exactContextNeedsRuntimeLookup = false;
+                    pResult->codePointerOrStubLookup.constLookup = CreateConstLookupToSymbol(_compilation.NodeFactory.ExternSymbol("NYI_LDVIRTFTN"));
+                }
+                else
+                {
+                    pResult->exactContextNeedsRuntimeLookup = false;
+                    targetMethod = targetMethod.GetCanonMethodTarget(CanonicalFormKind.Specific);
+
+                    // Get the slot defining method to make sure our virtual method use tracking gets this right.
+                    // For normal C# code the targetMethod will always be newslot.
+                    MethodDesc slotDefiningMethod = targetMethod.IsNewSlot ?
+                        targetMethod : MetadataVirtualMethodAlgorithm.FindSlotDefiningMethodForVirtualMethod(targetMethod);
+
+                    pResult->codePointerOrStubLookup.constLookup =
+                        CreateConstLookupToSymbol(
+                            _compilation.NodeFactory.ReadyToRunHelper(helperId, slotDefiningMethod));
+                }
+
+                // The current CoreRT ReadyToRun helpers do not handle null thisptr - ask the JIT to emit explicit null checks
+                // TODO: Optimize this
+                pResult->nullInstanceCheck = true;
+            }
+
+            pResult->hMethod = ObjectToHandle(targetMethod);
+
+            pResult->accessAllowed = CorInfoIsAccessAllowedResult.CORINFO_ACCESS_ALLOWED;
+
+            // We're pretty much done at this point.  Let's grab the rest of the information that the jit is going to
+            // need.
+            pResult->classFlags = getClassAttribsInternal(targetMethod.OwningType);
+
+            pResult->methodFlags = getMethodAttribsInternal(targetMethod);
+            Get_CORINFO_SIG_INFO(targetMethod, &pResult->sig, targetIsFatFunctionPointer);
+
+            if ((flags & CORINFO_CALLINFO_FLAGS.CORINFO_CALLINFO_VERIFICATION) != 0)
+            {
+                if (pResult->hMethod != pResolvedToken.hMethod)
+                {
+                    pResult->verMethodFlags = getMethodAttribsInternal(targetMethod);
+                    Get_CORINFO_SIG_INFO(targetMethod, &pResult->verSig);
+                }
+                else
+                {
+                    pResult->verMethodFlags = pResult->methodFlags;
+                    pResult->verSig = pResult->sig;
+                }
+            }
+
+            pResult->_secureDelegateInvoke = 0;
         }
     }
 }
